server/requirements.txt
Flask==3.1.0
flask-cors==4.0.1
fer==22.5.1
opencv-python==4.11.0.86
numpy==1.26.4
gunicorn==20.1.0
moviepy==1.0.3
tensorflow==2.15.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0